move fp32 context to utils
